refactor: HOC moved to hocs
