+++ packages/client/src/test/__mocks__/svgTransform.cjs
@@ -1,8 +1,8 @@
 module.exports = {
   process() {
-    return { code: 'module.exports = {};' };
+    return { code: 'module.exports = {};' }
   },
   getCacheKey() {
     return 'svgTransform';
   },
-};
+}
